Reuse URL validation
