+++ packages/ciphernode/config/src/app_config.rs
@@ -1,4 +1,7 @@
 use alloy::primitives::Address;
+use anyhow::anyhow;
+use anyhow::bail;
+use anyhow::Context;
 use anyhow::Result;
 use figment::{
     providers::{Format, Serialized, Yaml},
@@ -9,6 +12,7 @@ use std::{
     env,
     path::{Path, PathBuf},
 };
+use url::Url;
 
 #[derive(Debug, Deserialize, Serialize, PartialEq)]
 #[serde(untagged)]
@@ -18,6 +22,61 @@ pub enum Contract {
         deploy_block: Option<u64>,
     },
     AddressOnly(String),
+}
+
+#[derive(Clone)]
+pub enum RPC {
+    Http(String),
+    Https(String),
+    Ws(String),
+    Wss(String),
+}
+
+impl RPC {
+    pub fn from_url(url: &str) -> Result<Self> {
+        let parsed = Url::parse(url).context("Invalid URL format")?;
+        match parsed.scheme() {
+            "http" => Ok(RPC::Http(url.to_string())),
+            "https" => Ok(RPC::Https(url.to_string())),
+            "ws" => Ok(RPC::Ws(url.to_string())),
+            "wss" => Ok(RPC::Wss(url.to_string())),
+            _ => bail!("Invalid protocol. Expected: http://, https://, ws://, wss://"),
+        }
+    }
+
+    pub fn as_http_url(&self) -> String {
+        match self {
+            RPC::Http(url) | RPC::Https(url) => url.clone(),
+            RPC::Ws(url) | RPC::Wss(url) => {
+                let mut parsed = Url::parse(url).expect(&format!("Failed to parse URL: {}", url));
+                parsed
+                    .set_scheme(if self.is_secure() { "https" } else { "http" })
+                    .expect("http(s) are valid schemes");
+                parsed.to_string()
+            }
+        }
+    }
+
+    pub fn as_ws_url(&self) -> String {
+        match self {
+            RPC::Ws(url) | RPC::Wss(url) => url.clone(),
+            RPC::Http(url) | RPC::Https(url) => {
+                let mut parsed = Url::parse(url).expect(&format!("Failed to parse URL: {}", url));
+                parsed
+                    .set_scheme(if self.is_secure() { "wss" } else { "ws" })
+                    .expect("ws(s) are valid schemes");
+                parsed.to_string()
+            }
+        }
+    }
+
+    pub fn is_websocket(&self) -> bool {
+        matches!(self, RPC::Ws(_) | RPC::Wss(_))
+    }
+
+    pub fn is_secure(&self) -> bool {
+        matches!(self, RPC::Https(_) | RPC::Wss(_))
+    }
 }
 
 impl Contract {
@@ -63,10 +122,17 @@ impl Default for RpcAuth {
 pub struct ChainConfig {
     pub enabled: Option<bool>,
     pub name: String,
-    pub rpc_url: String, // We may need multiple per chain for redundancy at a later point
+    rpc_url: String, // We may need multiple per chain for redundancy at a later point
     #[serde(default)]
     pub rpc_auth: RpcAuth,
     pub contracts: ContractAddresses,
+}
+
+impl ChainConfig {
+    pub fn rpc_url(&self) -> Result<RPC> {
+        Ok(RPC::from_url(&self.rpc_url)
+            .map_err(|e| anyhow!("Failed to parse RPC URL for chain {}: {}", self.name, e))?)
+    }
 }
 
 #[derive(Debug, Deserialize, Serialize)]
